refactor(listener): extract channel commands
